requirements.txt
streamlit==1.28.0
langchain==0.0.350
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.35.0
pydantic==2.4.2
jinja2==3.1.2
weasyprint==60.2
openai==1.3.0
python-dotenv==1.0.0
numpy==1.24.3
